# backend/src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "PollSnap API",
    "description": "PollSnap - A polling platform API",
    "version": "1.0.0"
  },
  "host": "localhost:8888",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "tags": [
    { "name": "Health", "description": "Health check endpoints" },
    { "name": "Users", "description": "User management endpoints" },
    { "name": "Polls", "description": "Poll management endpoints" },
    { "name": "Questions", "description": "Question management endpoints" },
    { "name": "Options", "description": "Option management endpoints" },
    { "name": "Responses", "description": "Response submission endpoints" },
    { "name": "Analytics", "description": "Analytics and results endpoints" },
    { "name": "Dashboard", "description": "Dashboard endpoints" }
  ],
  "paths": {
    "/health/": {
      "get": {
        "tags": ["Health"],
        "description": "Basic health check",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/health/detailed": {
      "get": {
        "tags": ["Health"],
        "description": "Detailed health check",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/dashboard/stats": {
      "get": {
        "tags": ["Dashboard"],
        "description": "Get dashboard stats",
        "responses": { "default": { "description": "" } }
      }
    },
    "/v1/dashboard/trends": {
      "get": {
        "tags": ["Dashboard"],
        "description": "Get dashboard trends",
        "parameters": [
          { "name": "days", "in": "query", "required": false, "type": "integer" }
        ],
        "responses": { "default": { "description": "" } }
      }
    },
    "/v1/dashboard/recent-activity": {
      "get": {
        "tags": ["Dashboard"],
        "description": "Get recent activity",
        "responses": { "default": { "description": "" } }
      }
    },
    "/v1/dashboard/audience-insights": {
      "get": {
        "tags": ["Dashboard"],
        "description": "Get audience insights",
        "responses": { "default": { "description": "" } }
      }
    },
    "/v1/dashboard/plan-usage": {
      "get": {
        "tags": ["Dashboard"],
        "description": "Get plan usage",
        "responses": { "default": { "description": "" } }
      }
    },
    "/v1/users/me": {
      "get": {
        "tags": ["Users"],
        "description": "Get current user",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/polls/": {
      "get": {
        "tags": ["Polls"],
        "description": "List user's polls",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": ["Polls"],
        "description": "Create a new poll",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/polls/{id}": {
      "get": {
        "tags": ["Polls"],
        "description": "Get poll by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": ["Polls"],
        "description": "Update poll",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": ["Polls"],
        "description": "Delete poll",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/polls/{id}/publish": {
      "patch": {
        "tags": ["Polls"],
        "description": "Publish poll",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/polls/{id}/activate": {
      "post": {
        "tags": ["Polls"],
        "description": "Activate poll. Requires at least one question with 2+ options and an expiry date.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/polls/{id}/close": {
      "post": {
        "tags": ["Polls"],
        "description": "Close poll",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/polls/{id}/questions/": {
      "get": {
        "tags": ["Questions"],
        "description": "List questions for a poll",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": ["Questions"],
        "description": "Create a question",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/polls/{id}/questions/reorder": {
      "post": {
        "tags": ["Questions"],
        "description": "Reorder questions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/polls/{id}/questions/{questionId}": {
      "patch": {
        "tags": ["Questions"],
        "description": "Update question",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": ["Questions"],
        "description": "Delete question",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/polls/{id}/questions/{questionId}/options": {
      "get": {
        "tags": ["Options"],
        "description": "List options for a question",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": ["Options"],
        "description": "Create an option",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/polls/{id}/questions/{questionId}/options/reorder": {
      "post": {
        "tags": ["Options"],
        "description": "Reorder options",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/polls/{id}/questions/{questionId}/options/{optionId}": {
      "patch": {
        "tags": ["Options"],
        "description": "Update option",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "optionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": ["Options"],
        "description": "Delete option",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "optionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/polls/{id}/responses": {
      "post": {
        "tags": ["Responses"],
        "description": "Submit response to a poll",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/polls/{id}/analytics": {
      "get": {
        "tags": ["Analytics"],
        "description": "Get poll analytics (creator only)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Poll analytics data including response stats, question results, recent activity, and audience insights",
            "schema": {
              "type": "object",
              "properties": {
                "poll": { "type": "object" },
                "totalResponses": { "type": "number" },
                "goalProgress": { "type": "number", "nullable": true },
                "completionRate": { "type": "number" },
                "questions": { "type": "array", "items": { "type": "object" } },
                "recentVotes": { "type": "array", "items": { "type": "object" } },
                "audience": {
                  "type": "object",
                  "properties": {
                    "geographic": { "type": "array", "items": { "type": "object" } },
                    "devices": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/polls/{id}/results": {
      "get": {
        "tags": ["Analytics"],
        "description": "Get public poll results (published polls only)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  }
}